Replace instrument_manager.py and settings.py with a module for storing persistent state

+++ ivtools/io.py
@@ -56,6 +56,9 @@ class MetaHandler(object):
             self.prettykeys = []
             self.moduledir = os.path.split(__file__)[0]
 
+    def __repr__(self):
+        #return self.meta.__repr__()
+        return pd.Series({**self.meta, **self.static}).__repr__()
 
     def self_merge(self, columns=None):
         '''
@@ -745,12 +748,13 @@ def write_meta_csv(data, filepath):
     ''' Write the non-array data to a text file.  Only first row of dataframe considered!'''
     dtype = type(data)
     if dtype is pd.Series:
-        s = pd.read_pickle(pjoin(root, f))
+        #s = pd.read_pickle(pjoin(root, f))
+        s = data
     elif dtype is pd.DataFrame:
         # Only save first row metadata -- Usually it's the same for all
-        df = pd.read_pickle(pjoin(root, f))
-        s = df.iloc[0]
-        s['nloops'] = len(df)
+        #df = pd.read_pickle(pjoin(root, f))
+        s = data.iloc[0]
+        s['nloops'] = len(data)
     elif dtype is list:
         s = pd.Series(data[0])
     elif dtype is dict:
@@ -770,7 +774,7 @@ def set_readonly(filepath):
     from stat import S_IREAD, S_IRGRP, S_IROTH
     os.chmod(filepath, S_IREAD|S_IRGRP|S_IROTH)
 
-def plot_datafiles(datadir, maxloops=500, x='V', y='I', smoothpercent=1):
+def plot_datafiles(datadir, maxloops=500, x='V', y='I', smoothpercent=1, overwrite=False):
    # Make a plot of all the .s and .df files in a directory
    # Save as pngs with the same name
    # TODO: Optionally group by sample, making one plot per sample
@@ -781,34 +785,36 @@ def plot_datafiles(datadir, maxloops=500, x='V', y='I', smoothpercent=1):
    fig, ax = plt.subplots()
 
    for sfn in series_fns:
-      s = pd.read_pickle(sfn)
-      s.I *= 1e6
-      s.units['I'] = '$\mu$A'
-      smoothn = max(int(smoothpercent * len(s.V) / 100), 1)
-      ivplot.plotiv(analyze.moving_avg(s, smoothn, columns=None), x=x, y=y, ax=ax)
       pngfn = sfn[:-2] + '.png'
       pngfp = os.path.join(datadir, pngfn)
-      if 'width_nm' in s:
-          plt.title('{}, Width={}nm, Thickness={}nm'.format(s['layer_1'], s['width_nm'], s['thickness_1']))
-      plt.savefig(pngfp)
-      print('Wrote {}'.format(pngfp))
-      ax.cla()
+      if overwrite or not os.path.isfile(pngfp):
+        s = pd.read_pickle(sfn)
+        s.I *= 1e6
+        s.units['I'] = '$\mu$A'
+        smoothn = max(int(smoothpercent * len(s.V) / 100), 1)
+        ivplot.plotiv(analyze.moving_avg(s, smoothn, columns=None), x=x, y=y, ax=ax)
+        if 'width_nm' in s:
+            plt.title('{}, Width={}nm, Thickness={}nm'.format(s['layer_1'], s['width_nm'], s['thickness_1']))
+        plt.savefig(pngfp)
+        print('Wrote {}'.format(pngfp))
+        ax.cla()
 
    for dffn in dataframe_fns:
-      df = pd.read_pickle(dffn)
-      df.I *= 1e6
-      df['units'] = len(df) * [{'V':'V', 'I':'$\mu$A'}]
-      step = int(ceil(len(df) / maxloops))
-      smoothn = max(int(smoothpercent * len(df.iloc[0].V) / 100), 1)
-      ivplot.plotiv(analyze.moving_avg(df[::step], smoothn), alpha=.6, ax=ax)
       pngfn = dffn[:-3] + '.png'
       pngfp = os.path.join(datadir, pngfn)
-      s = df.iloc[0]
-      if 'width_nm' in s:
-          plt.title('{}, Width={}nm, Thickness={}nm'.format(s['layer_1'], s['width_nm'], s['thickness_1']))
-      plt.savefig(pngfp)
-      print('Wrote {}'.format(pngfp))
-      ax.cla()
+      if overwrite or not os.path.isfile(pngfp):
+        df = pd.read_pickle(dffn)
+        df.I *= 1e6
+        df['units'] = len(df) * [{'V':'V', 'I':'$\mu$A'}]
+        step = int(ceil(len(df) / maxloops))
+        smoothn = max(int(smoothpercent * len(df.iloc[0].V) / 100), 1)
+        ivplot.plotiv(analyze.moving_avg(df[::step], smoothn), alpha=.6, ax=ax)
+        s = df.iloc[0]
+        if 'width_nm' in s:
+            plt.title('{}, Width={}nm, Thickness={}nm'.format(s['layer_1'], s['width_nm'], s['thickness_1']))
+        plt.savefig(pngfp)
+        print('Wrote {}'.format(pngfp))
+        ax.cla()
 
    plt.close(fig)
